requirements.txt
streamlit>=1.36
pandas>=2.2
python-dotenv>=1.0
eth-account>=0.13
google-generativeai>=0.7
openai>=0.27.8
requests>=2.31